Revert rename of spyderlib.widgets.varexp to spyderlib.widgets.editors
This reverts commit 562464e1cc9c7c6e8c734f65803286d84a40d557.
