+++ spyderlib/widgets/varexp/arrayeditor.py
