+++ spyderlib/widgets/varexp/texteditor.py
